refactor: 컴포넌트별 파일 분리 및 렌더링 로직 개선
main.basic.js 파일의 컴포넌트들을 각각의 파일로 분리하여 모듈성을 높이고, 렌더링 로직을 별도의 파일로 분리하여 코드 구조를
  개선했습니다. 이를 통해 코드의 가독성과 유지보수성이 향상되었습니다.

  주요 변경 사항:

   * src/basic/components 폴더를 생성하고, 각 UI 컴포넌트(Header, Layout, Selector, CartItem, HelpModal)를 별도의 파일로 분리했습니다.
   * 렌더링 관련 함수(renderProductOptions, render)를 src/basic/render.js 파일로 분리했습니다.
   * main.basic.js는 각 컴포넌트와 렌더링 함수를 직접 import 하여 사용하도록 수정했습니다.
   * 더 이상 사용되지 않는 src/basic/dom.js 파일을 삭제했습니다.
   * 모듈 분리에 따른 import 경로를 수정했습니다.
