+++ src/basic/components/Layout.js
@@ -1,0 +1,34 @@
+export function createLayout() {
+    const grid = document.createElement('div');
+    grid.className = 'grid grid-cols-1 lg:grid-cols-[1fr_360px] gap-6 flex-1 overflow-hidden';
+
+    const left = document.createElement('div');
+    left.className = 'bg-white border border-gray-200 p-8 overflow-y-auto';
+    const right = document.createElement('div');
+    right.className = 'bg-black text-white p-8 flex flex-col';
+    right.innerHTML = `
+<h2 class="text-xs font-medium mb-5 tracking-extra-wide uppercase">Order Summary</h2>
+<div class="flex-1 flex flex-col">
+  <div id="summary-details" class="space-y-3"></div>
+  <div class="mt-auto">
+    <div id="discount-info" class="mb-4"></div>
+    <div id="cart-total" class="pt-5 border-t border-white/10">
+      <div class="flex justify-between items-baseline">
+        <span class="text-sm uppercase tracking-wider">Total</span>
+        <div class="text-2xl tracking-tight">₩0</div>
+      </div>
+      <div id="loyalty-points" class="text-xs text-blue-400 mt-2 text-right" style="display:none">적립 포인트: 0p</div>
+    </div>
+    <div id="tuesday-special" class="mt-4 p-3 bg-white/10 rounded-lg hidden">
+      <div class="flex items-center gap-2">
+        <span class="text-2xs">🎉</span>
+        <span class="text-xs uppercase tracking-wide">Tuesday Special 10% Applied</span>
+      </div>
+    </div>
+  </div>
+</div>
+<button class="w-full py-4 bg-white text-black text-sm font-normal uppercase tracking-super-wide cursor-pointer mt-6">Proceed to Checkout</button>
+<p class="mt-4 text-2xs text-white/60 text-center leading-relaxed">Free shipping on all orders.<br><span id="points-notice">Earn loyalty points with purchase.</span></p>
+`;
+    return { grid, left, right };
+}
